requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
autoviz==0.1.905
bcrypt==4.1.3
bleach==6.1.0
bokeh==3.4.3
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.12.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fonttools==4.53.1
fpdf==1.7.2
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.1
grpcio-status==1.62.2
h11==0.14.0
holoviews==1.19.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
hvplot==0.10.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jose==1.0.0
jwt==1.3.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.4
openai==1.36.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pandas-dq==1.29
panel==1.4.4
param==2.1.1
parso==0.8.4
passlib==1.7.4
patsy==0.5.6
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pure_eval==0.2.3
pyamg==5.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pyviz_comms==3.0.2
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlmodel==0.0.19
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.19.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wordcloud==1.9.3
xgboost==1.6.2
xlrd==2.0.1
xyzservices==2024.6.0
